requirements.txt
dj-database-url==0.4.2
Django==1.11
django-redis==4.8.0
django-redis-cache==1.7.1
djangorestframework==3.7.7
gunicorn==19.7.1
psycopg2==2.7.3.2
pytz==2018.3
redis==2.10.6
whitenoise==3.3.1
